Refactor and restructure part 3
