+++ tests/PSRule.Rules.Azure.Tests/DependencyMapTests.cs
@@ -4,7 +4,6 @@
 using System.Linq;
 using Newtonsoft.Json.Linq;
 using PSRule.Rules.Azure.Arm.Deployments;
-using PSRule.Rules.Azure.Data.Template;
 using static PSRule.Rules.Azure.Arm.Deployments.DeploymentVisitor;
 
 namespace PSRule.Rules.Azure
